Buying goods done.

+++ src/com/codecool/api/HardWareStore.java
@@ -13,7 +13,7 @@ public class HardWareStore extends Store {
     
     private List<BoughtComponent> stock = new ArrayList<>();
     private double money;
-    private int number;
+    private double number;
     
     public HardWareStore(int money) {
         this.stock = stock;
@@ -29,11 +29,11 @@ public class HardWareStore extends Store {
         money += value;
     }
     
-    public int getNumber() {
+    public double getNumber() {
         return number;
     }
     
-    public void setNumber(int number) {
+    public void setNumber(double number) {
         this.number += number;
     }
     
@@ -43,7 +43,7 @@ public class HardWareStore extends Store {
         return stock;
     }
     
-    public void addComponent(Components component, int number) {
+    public void addComponent(Components component, double number) {
         BoughtComponent toStockComponent = new BoughtComponent(component.getName(), number, component);
         stock.add(toStockComponent);
     }
